requirements.txt
langchain>=0.0.339
langchain-deepseek>=0.0.3
pydantic>=1.10.8
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
numpy>=1.24.3
alpaca-py>=0.8.2
faiss-cpu>=1.7.4
chromadb>=0.4.18
pandas>=2.0.3
tiktoken>=0.5.1
sec-api>=1.0.17
fastapi>=0.104.0
uvicorn>=0.23.2
langchain_openai
openai
langchain-community
sentence-transformers
langchain-huggingface
transformers
yfinance
stockstats
torch
matplotlib
python-multipart
langchain_huggingface